app: src: modules: Add battery module
Convert the sampler module to a battery module.
The battery module uses the nRF Fuel Gage library and sensor API to
readout information about the lp803448 thingy91x battery from the
npm1300.

Currently the fuel gage parameters are only printed.
Encoding of the battery information is not implemented yet, as well as
publication on the PAYLOAD channel.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>