app: location: Use zbus subscriber instead of listener
Updated the location module to use zbus subscribe instead of a listener
for consistency and easier logic. Removed semaphores and modem lib
connection handling. Connection handling might be needed unless the
location lib doesn't handle this internally, but it should probably be
dealt with using the Zephyr State Machine Framework.

Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>