workflow: Add workflow to attach release assets
When a release is created this new workflow calls the other build
and test workflows and uploads the tested artifacts into github
release.
The existing workflows were modified to make them callable from other
workflows.
Build workflow now uploads the artifacts so that they can be added to
the release. The build workflow uploads the hex, bin and the .config
file. The names of the files will have the sha if built from main and
the tag name if built from a tag.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>