workflow: on_target: Reuse build workflow to build firmware
Previously the on_target workflow was re-doing the steps done by the
build workflow. But since now the build workflow uploads the built
firmware as artifacts, we now have the possibility to re-use it. The
on_target workflow now calls the build worfklow and downloads its
artifaacts to perform the tests.

The test fraemwork code was modified to accept new input argument
firmware-hex that points to the hex file to test.
Also renamed the python file to have the test_ prefix so that pytest
auto-detects it when invoked without any options.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>