workflow: attach_release_assets: Fix publishing to nrfcloud
A repo checkout was needed in order to make the gh command succeed. Also
the version is now obtained from github event instead of depending on
the environment variable. The environment variable was previously set by
build workflow which is now triggered by the on_target workflow. The
added link the chain makes the environment variable inaccesible to this
workflow.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>