modules: app: Convert to MSG subscriber
Convert to MSG subscriber to ensure that the app module is
able to receive messages from other modules with guaranteed
delivery.

Fixes issue where the trigger module would send a POLL and DATA SAMPLE
trigger at the same time, causing the app module to miss the POLL due
to the payload being overwritten.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>