app: transport: Fix watchdog timeout on LTE connection wait
Moved the task_wdt_add function after the point where time is obtained
from network. This is because it was observed that network connection
sometimes take longer than the watchdog timeout. This leads to
unnecessary watchdog timeout and restarts. Adding the task watchdog
after network time is obtained is better because we only want to capture
any abnormal hanging in the thread's main loop anyway.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>