modules: network: Add Energy Estimate reporting
The network module now reports energy estimate periodically to the
Connection Information LWM2M object.

To achieve this, the network module is now made to listen to the trigger
channel and upon receiving a trigger, it retrieves the energy estimate
using the link control API and encodes it into cbor data and sends it to
the payload channel. Tested and found to work with hello nrfcloud.

Unit test added.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>