modules: trigger: Fix entering frequent poll state
Removed UNBLOCKED state.
Statemachine reworked to make the FREQUENT_POLL and NORMAL states
as direct children of CONNECTED state (with BLOCKED as sibling).
TheUNBLOCKED state was automatically transitioning into
FREQUENT_POLL state irrespective of what the old state. This lead to the
system entering the frequent polling state when GNSS was enabled and
therefore there was a transition from BLOCKED to UNBLOCKED state.

Fixes CIA-1295

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>