trigger: Fix frequent poll duration timer related bugs
- Fix a bug where the frequent poll duration timer was not stopped on
disconnect event. This lead to fault transitions when it expired.
- Fix a bug where when the frequent poll duration timer expires in
blocked state, there will be a sample trigger sent. This was because
the state machine was transitioning to normal state and executing normal
mode entry functions. This is now fixed by only changing the trigger
mode in the state object but staying in the blocked state.

Add native_sim based unit test for trigger module. Covers most of the
state transitions. Fixed a few bugs discovered in the trigger module.

Not tested yet:
- Button press in blocked state
- config rcvd in blocked state
- Button press in frequent poll state
- config rcvd in frequent poll state

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>