boards: thingy91x_nrf9151_ns: Revert to low power mode
Revert to low power mode to sample valid environmental data
after boot.

In ultra low power mode the BSEC driver doesn't give valid
values before the ULP sample time has surpassed. (5 minutes)

Ideally the BSEC driver should do an inital sample right after boot,
but this is not currently the case.

This revert will cause the firmware to draw more power, but is a needed
change given the fact that the firmware is an OOB demo.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>