location: Increase location work queue's stack size
We see that the location_core module misbehaves on two devices where it
sends corrupted location data to the coap library and the encoding
fails. And this happens, seemingly, in a loop which causes several
updates being sent to the network in short interval. This inturn leads
to the zbus queue overfilling and finally an assert.  We suspect a
possible stack overflow that may lead to this. And therefore we increase
the stack size of the location workqueue to 8k as in other NCS samples.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>