modules: trigger: Don't schedule triggers immediately in normal mode
When entering normal mode, schedule triggers after the respective
timeouts, not immediately.

This fix prevents the trigger module from firing triggers every time
the GNSS is disabled in normal mode.
This issue occurs when a trigger is sent, and GNSS finishes sampling,
causing an endless loop of back-and-forth triggers instead of
returning to the expected interval.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>