modules: trigger: Update trigger module
When entering normal mode, schedule triggers after the respective
timeouts, not immediately.
This fix prevents the trigger module from firing triggers every time
the GNSS is disabled in normal mode.
This issue occurs when a trigger is sent, and GNSS finishes sampling,
causing an endless loop of back-and-forth triggers instead of
returning to the expected interval.

Add separate trigger for shadow polling. Polling now occurs
every 20 seconds when the trigger module is actively polling cloud,
and data sampling every 60 seconds. This is to preserve power and
data usage.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>