manifest: Update sdk-nrf
Update sdk-nrf to pull in change that limits getting modem
parameters to when the modem is not in offline mode.

The on target test currently fails quite often due to the modem
being put in offline mode in between tests.
This in turn triggers a warning in the memfault lte integration layer
that calls the modem info API that fail.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>