on_target: Add test case for network loss
Add test case for network loss
Since we are explicitly setting the mode in offline mode, also
update calling of lte lc coneval API to not send fatal error
if called with an unsupported functional mode.

Also, refactor the test a bit to make it easier to add new testcases.
and update wait_for_str test function to expect the order to the input
list.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>