modules: location: Send LOCATION_SEARCH_DONE on error
Send LOCATION_SEARCH_DONE event on LOCATION_SEARCH_DONE so that the
trigger modules gets out of the blocked state on error condition and
tries again during the next location update. This situation happens when
the network connectivity was lost and then was restored again and the
location module tries to fetch the location and fails due to coap
issues.

Signed-off-by: Balaji Srinivasan <balaji.srinivasan@nordicsemi.no>