modules: Update buffer size and name
Rename 'string' payload message type to 'buffer'.

The payload exclusively carries CBOR-encoded data, making 'buffer' a
more accurate term than 'string' since the data is
not treated as a string.

Also increase the size of the buffer as we have seen indications that
in might be too small in some cases.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>