modules: Unify timeout for `zbus_chan_pub`
Use 1 second as the default timeout when publishing on a zbus channel.
Previously, different values were used, sometimes `K_NO_WAIT`. Using
`K_NO_WAIT` is problematic if the channel is locked by another context
when calling `zbus_chan_pub`. If this occurs, the API fails, which we
currently handle by asserting. Increasing the timeout increases the
chance of the API succeeding.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>