workflows: update pmr_nrf91.txt for external flash partitions
The partition named external_flash in the external_flash region should
be placed by partition manager to fill the remainder of the external
flash. This allows adding dynamic partitions in some build
configurations. The static partition map was updated in sdk-nrf, so we
need to update the partition map we compare against in CI accordingly.

Signed-off-by: Gregers Gram Rygg <gregers.gram.rygg@nordicsemi.no>