memfault: Clear modem traces after boot on no valid coredump
Clear modem traces after boot on no valid coredump.
This is to prevent any issues related to non-volitile variables
in the modem trace flash backend such as offset and size.

We have observed that some traces are corrupted and to potentially
improve this we clear the modem traces after boot if there is
no valid coredump.

Increase timeout when crashing to 10 seconds to allow more time for
memfault to save the coredump and modem traces to be saved to flash.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>